+++ alma/controllers/hook/StateHookController.php
@@ -31,14 +31,33 @@ if (!defined('_PS_VERSION_')) {
 use Alma\API\Client;
 use Alma\API\RequestError;
 use Alma\PrestaShop\Helpers\ClientHelper;
+use Alma\PrestaShop\Helpers\ConfigurationHelper;
 use Alma\PrestaShop\Helpers\OrderHelper;
 use Alma\PrestaShop\Helpers\SettingsHelper;
+use Alma\PrestaShop\Helpers\ShopHelper;
 use Alma\PrestaShop\Hooks\AdminHookController;
 use Alma\PrestaShop\Logger;
 use Alma\PrestaShop\Model\OrderData;
 
 final class StateHookController extends AdminHookController
 {
+    /**
+     * @var SettingsHelper
+     */
+    protected $settingsHelper;
+
+    /**
+     * HookController constructor.
+     *
+     * @param $module Alma
+     */
+    public function __construct($module)
+    {
+        parent::__construct($module);
+
+        $this->settingsHelper = new SettingsHelper(new ShopHelper(), new ConfigurationHelper());
+    }
+
     /**
      * Checks if user is logged in as Employee or is an API Webservice call
      *
@@ -92,7 +111,7 @@ final class StateHookController extends AdminHookController
                 }
                 break;
             case $idStatePaymentTrigger:
-                if (SettingsHelper::isPaymentTriggerEnabledByState()) {
+                if ($this->settingsHelper->isPaymentTriggerEnabledByState()) {
                     $this->triggerPayment($alma, $idPayment, $order);
                 }
                 break;