+++ alma/lib/Helpers/ConfigurationHelper.php
@@ -1,5 +1,6 @@
-{*
- * 2018-2023 Alma SAS
+<?php
+/**
+ * 2018-2023 Alma SAS.
  *
  * THE MIT LICENSE
  *
@@ -19,16 +20,48 @@
  * @author    Alma SAS <contact@getalma.eu>
  * @copyright 2018-2023 Alma SAS
  * @license   https://opensource.org/licenses/MIT The MIT License
- *}
-<form id="alma-inpage-plan-{$keyPlan|escape:'htmlall':'UTF-8'}" class="alma-inpage"
-      data-action="{$action}"
-      data-apimode="{$apiMode|escape:'htmlall':'UTF-8'}"
-      data-merchantid="{$merchantId|escape:'htmlall':'UTF-8'}"
-      data-isinpageenabled="{$isInPageEnabled|escape:'htmlall':'UTF-8'}"
-      data-installment="{$installment|escape:'htmlall':'UTF-8'}"
-      data-deferreddays="{$deferredDays|escape:'htmlall':'UTF-8'}"
-      data-deferredmonths="{$deferredMonths|escape:'htmlall':'UTF-8'}"
-      data-purchaseamount="{$creditInfo.totalCart|escape:'htmlall':'UTF-8'}"
-      data-locale="{$locale|escape:'htmlall':'UTF-8'}">
-    <div id="alma-inpage-iframe-plan-{$keyPlan|escape:'htmlall':'UTF-8'}" class="alma-inpage-iframe"></div>
-</form>
+ */
+
+namespace Alma\PrestaShop\Helpers;
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+/**
+ * Class LanguageHelper.
+ */
+class ConfigurationHelper
+{
+    /**
+     * Get a single configuration value (in one language only).
+     *
+     * @param string $key Key wanted
+     * @param int $idLang Language ID
+     *
+     * @codeCoverageIgnore
+     *
+     * @return string|false Value
+     */
+    public function get($key, $idLang = null, $idShopGroup = null, $idShop = null, $default = false)
+    {
+        return \Configuration::get($key, $idLang, $idShopGroup, $idShop, $default);
+    }
+
+    /**
+     * Check if key exists in configuration.
+     *
+     * @param string $key
+     * @param int $idLang
+     * @param int $idShopGroup
+     * @param int $idShop
+     *
+     * @codeCoverageIgnore
+     *
+     * @return bool
+     */
+    public function hasKey($key, $idLang = null, $idShopGroup = null, $idShop = null)
+    {
+        return \Configuration::hasKey($key, $idLang, $idShopGroup, $idShop);
+    }
+}